pyproject.toml
[tool.poetry]
name = "game-code-generator"
version = "0.1.0"
description = "AI-powered game development system using Phaser"
authors = ["Subir Mansukhani <subir.m@gmail.com>"]
readme = "README.md"
packages = [{include = "game_generator"}]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
watchdog = "^6.0.0"
rich = "^13.9.0"
python-dotenv = "1.0.1"
anthropic = "0.39.0"
pygame = "2.6.1"
numpy = "2.1.3"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
